refactor: Rename namespace for DataProvider feature to `Sample`
